service: factor out detail endpoint from engagements
The code really didn't belong there; it just started out as handling
engagements only.
